Converted index.ts files to only list exports
