+++ packages/common/src/index.ts
@@ -14,96 +14,14 @@
  * limitations under the License.
  */
 
-import * as auth from "@esri/arcgis-rest-auth";
+/**
+ * Provides common functions.
+ *
+ * @module common
+ */
 
 export * from "./generalHelpers";
+export * from "./interfaces";
+export * from "./resourceHelpers";
 export * from "./restHelpers";
 export * from "./templatization";
-
-// ------------------------------------------------------------------------------------------------------------------ //
-
-/**
- * A solution template AGO item
- */
-export interface ISolutionItem {
-  /**
-   * Item base section JSON
-   */
-  item: any;
-  /**
-   * Item data section JSON
-   */
-  data: ISolutionItemData;
-}
-
-export interface ISolutionItemData {
-  /**
-   * General information about the solution template
-   */
-  metadata: any;
-  /**
-   * The collection of templates
-   */
-  templates: IItemTemplate[]
-}
-
-export interface IItemTemplate {
-  /**
-   * Item's AGO id
-   */
-  itemId: string;
-  /**
-   * AGO item type name
-   */
-  type: string;
-  /**
-   * Fairly unique identifier; set to 'i' + chars 2-8 of a random number in base 36
-   */
-  key: string;
-  /**
-   * Item base section JSON
-   */
-  item: any;
-  /**
-   * Item data section JSON
-   */
-  data: any;
-  /**
-   * References to item resources
-   */
-  resources: any[];
-  /**
-   * List of ids of AGO items needed by this item
-   */
-  dependencies: string[];
-  /**
-   * Miscellaneous item-specific properties
-   */
-  properties: any;
-  /**
-   * Estimated relative cost of deploying this item; corresponds to number of progressCallback
-   * function calls made during while deploying it
-   */
-  estimatedDeploymentCostFactor: number;
-}
-
-
-export interface IItemTemplateConversions {
-  convertItemToTemplate(
-    itemInfo: any,
-    userSession: auth.UserSession
-  ): Promise<IItemTemplate>;
-  createItemFromTemplate(
-    template: IItemTemplate,
-    templateDictionary: any,
-    userSession: auth.UserSession,
-    progressTickCallback: () => void
-  ): Promise<string>;
-}
-
-/**
- * Structure for mapping from item type to module with type-specific template-handling code
- */
-export interface IItemTypeModuleMap {
-  [itemType: string]: IItemTemplateConversions;
-}
