+++ packages/creator/src/index.ts
@@ -20,45 +20,5 @@
  * @module creator
  */
 
-import * as auth from "@esri/arcgis-rest-auth";
-import * as common from "@esri/solution-common";
-import * as createSolutionItem from "./createSolutionItem";
-import * as portal from "@esri/arcgis-rest-portal";
-
-// ------------------------------------------------------------------------------------------------------------------ //
-
-export interface IPortalSubset {
-  name: string;
-  id: string;
-  restUrl: string;
-  portalUrl: string;
-  urlKey: string;
-}
-
-export function createSolution(
-  groupId: string,
-  destUrl: string,
-  userSession: auth.UserSession
-): Promise<string> {
-  return new Promise<string>((resolve, reject) => {
-
-    // Fetch group item info
-
-    // Create an internal representation of the new solution item using group item info
-
-    // Fetch group contents
-
-    // For each group content item,
-    //   * fetch item & data infos
-    //   * create item & data JSONs
-    //   * extract dependency ids & add them into list of group contents
-    //   * templatize select components in item & data JSONs (e.g., extents)
-    //   * add JSONs into items list in solution item representation
-
-    // Create solution item using internal representation & and the data JSON
-
-    resolve("createSolution");
-  });
-}
-
-// ------------------------------------------------------------------------------------------------------------------ //
+export * from "./creator";
+export * from "./createSolutionItem";
