+++ packages/feature-layer/src/index.ts
@@ -20,59 +20,4 @@
  * @module feature-layer
  */
 
-import * as auth from "@esri/arcgis-rest-auth";
-import * as common from "@esri/solution-common";
-
-// ------------------------------------------------------------------------------------------------------------------ //
-
-export function convertItemToTemplate(
-  itemInfo: any,
-  userSession: auth.UserSession
-): Promise<common.IItemTemplate> {
-  return new Promise<common.IItemTemplate>(resolve => {
-    console.log("convertItemToTemplate for a feature-layer");
-    resolve(undefined);
-  });
-}
-
-export function createItemFromTemplate(
-  template: common.IItemTemplate,
-  templateDictionary: any,
-  userSession: auth.UserSession,
-  progressTickCallback: () => void
-): Promise<string> {
-  return new Promise<string>((resolve, reject) => {
-    console.log("createItemFromTemplate for a " + template.type + " (" + template.itemId + ")");
-
-    // Replace the templatized symbols in a copy of the template
-    let newItemTemplate = common.cloneObject(template) as common.IItemTemplate;
-    newItemTemplate = common.replaceInTemplate(newItemTemplate, templateDictionary);
-
-    // Create the item, then update its URL with its new id
-    common.createFeatureService(newItemTemplate.item, newItemTemplate.data,
-      {authentication: userSession}, templateDictionary.folderId)
-    .then(
-      createResponse => {
-        progressTickCallback();
-
-        if (createResponse.success) {
-          // Update the template dictionary with the new id & url
-          templateDictionary[template.itemId] = {
-            id: createResponse.serviceItemId,
-            url: createResponse.serviceurl
-          };
-
-          // Add the feature service's layers and tables to it
-          resolve(createResponse.serviceItemId);
-
-
-
-
-        } else {
-          reject(common.fail());
-        }
-      },
-      e => reject(common.fail(e))
-    );
-  });
-}
+export * from "./feature-layer";
