+++ packages/storymap/src/index.ts
@@ -20,60 +20,4 @@
  * @module storymap
  */
 
-import * as auth from "@esri/arcgis-rest-auth";
-import * as common from "@esri/solution-common";
-
-// ------------------------------------------------------------------------------------------------------------------ //
-
-export function convertItemToTemplate(
-  itemInfo: any,
-  userSession: auth.UserSession
-): Promise<common.IItemTemplate> {
-  return new Promise<common.IItemTemplate>(resolve => {
-    if (itemInfo.type === "StoryMap") {
-      console.log("convertItemToTemplate for a next-gen storymap");
-    } else {
-      console.log("convertItemToTemplate for a first-gen storymap");
-    }
-    resolve(undefined);
-  });
-}
-
-export function createItemFromTemplate(
-  template: common.IItemTemplate,
-  templateDictionary: any,
-  userSession: auth.UserSession,
-  progressTickCallback: () => void
-): Promise<string> {
-  return new Promise<string>((resolve, reject) => {
-    if (template.type === "StoryMap") {
-      console.log("createItemFromTemplate for a " + template.type + " (" + template.itemId + ")");
-    } else {
-      console.log("createItemFromTemplate for a " + template.type + " (StoryMap " + template.itemId + ")");
-    }
-    resolve("");
-  });
-}
-
-export function isAStoryMap(
-  template: common.IItemTemplate,
-): boolean {
-  const url = common.getProp(template, "item.url");
-  if (!url) {
-    return false;
-  } else if (template.type === "StoryMap") {
-    return true;
-  }
-
-  return [
-    /\/apps\/Cascade\//i,
-    /\/apps\/MapJournal\//i,
-    /\/apps\/MapSeries\//i,
-    /\/apps\/MapTour\//i,
-    /\/apps\/Shortlist\//i,
-    /\/apps\/StoryMap\//i,
-    /\/apps\/StoryMapBasic\//i,
-    /\/apps\/StorytellingSwipe\//i
-  ].some(pattern => pattern.test(url));
-}
-
+export * from "./storymap";
