+++ packages/viewer/src/viewer.ts
@@ -20,4 +20,20 @@
  * @module viewer
  */
 
-export * from "./viewer";
+import * as auth from "@esri/arcgis-rest-auth";
+import * as common from "@esri/solution-common";
+import * as portal from "@esri/arcgis-rest-portal";
+
+// ------------------------------------------------------------------------------------------------------------------ //
+
+export function getSolutionHierarchy(
+  groupId: string,
+  destUrl: string,
+  userSession: auth.UserSession
+): Promise<string> {
+  return new Promise<string>((resolve, reject) => {
+    resolve("getSolutionHierarchy");
+  });
+}
+
+// ------------------------------------------------------------------------------------------------------------------ //
